+++ resources/views/admin/profile/edit.blade.php
@@ -1,0 +1,27 @@
+@extends('admin.dashboard')
+@section('dashboard_content')
+    <div class="container">
+        <h2 class="fs-4 text-secondary my-4">
+            {{ __('Profile') }}
+        </h2>
+        <div class="card p-4 mb-4 bg-white shadow rounded-lg">
+
+            @include('admin.profile.partials.update-profile-information-form')
+
+        </div>
+
+        <div class="card p-4 mb-4 bg-white shadow rounded-lg">
+
+
+            @include('admin.profile.partials.update-password-form')
+
+        </div>
+
+        <div class="card p-4 mb-4 bg-white shadow rounded-lg">
+
+
+            @include('admin.profile.partials.delete-user-form')
+
+        </div>
+    </div>
+@endsection
