+++ resources/views/admin/profile/partials/delete-user-form.blade.php
@@ -11,12 +11,13 @@
 
     <!-- Modal trigger button -->
     <button type="button" class="btn btn-danger" data-bs-toggle="modal" data-bs-target="#delete-account">
-        {{__('Delete Account')}}
+        {{ __('Delete Account') }}
     </button>
 
     <!-- Modal Body -->
     <!-- if you want to close by clicking outside the modal, delete the last endpoint:data-bs-backdrop and data-bs-keyboard -->
-    <div class="modal fade" id="delete-account" tabindex="-1" data-bs-backdrop="static" data-bs-keyboard="false" role="dialog" aria-labelledby="delete-account" aria-hidden="true">
+    <div class="modal fade" id="delete-account" tabindex="-1" data-bs-backdrop="static" data-bs-keyboard="false"
+        role="dialog" aria-labelledby="delete-account" aria-hidden="true">
         <div class="modal-dialog modal-dialog-centered modal-lg" role="document">
             <div class="modal-content">
                 <div class="modal-header">
@@ -34,19 +35,20 @@
                 <div class="modal-footer">
                     <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Cancel</button>
 
-                    <form method="post" action="{{ route('profile.destroy') }}" class="p-6">
+                    <form method="post" action="{{ route('admin.profile.destroy', $data->id) }}" class="p-6">
                         @csrf
                         @method('delete')
 
 
                         <div class="input-group">
 
-                            <input id="password" name="password" type="password" class="form-control" placeholder="{{ __('Password') }}" />
+                            <input id="password" name="password" type="password" class="form-control"
+                                placeholder="{{ __('Password') }}" />
 
                             @error('password')
-                            <span class="invalid-feedback mt-2" role="alert">
-                                <strong>{{ $errors->userDeletion->get('password')}}</strong>
-                            </span>
+                                <span class="invalid-feedback mt-2" role="alert">
+                                    <strong>{{ $errors->userDeletion->get('password') }}</strong>
+                                </span>
                             @enderror
